Validate element when assert disabled/enabled
Add validation to check if the element is applicable for the
disabled/enabled condition. The `Condition.enabled/disabled` is calling
the Selenium's `WebElement.isEnabled()` that should be called on the
`input` element. Without knowing that enabled/disabled should be called
on the `input`, the Selenide's user can call on an any other element
that has `disabled` attribute, and it will either fail with an ambiguous
error message or even worse - will pass, which will be a false positive.
With this validation, the test will fail if called on a non-input
element. Thus, user will know exactly what went wrong.