set max upload size to protect from vulnerability CVE-2023-24998 in commons-fileupload
Quote:
Apache Commons FileUpload through 1.4 does not limit the number of request parts to be processed resulting in the possibility of an attacker triggering a DoS with a malicious upload or series of uploads. Note that, like all of the file upload limits, the new configuration option (FileUploadBase#setFileCountMax) is not enabled by default and must be explicitly configured. NOTE: Checkmarx strongly recommends our customers to update Apache Commons FileUpload to version 1.5 and enable the new configuration option (FileUploadBase#setFileCountMax). Updating to version 1.5 but not enabling the new configuration option will still leave you vulnerable to CVE-2023-24998.