#1636 collect browser logs using BiDi protocol
Currently, it's needed only for FireFox because the standard webdriver method `webDriver.manage().logs()` is not implemented in FF.

Implementation details:
In BiDi, we have to start listening for "Log.entryAdded" events immediately after webdriver start and collect these logs somewhere. The best place I managed to find was `WebDriverInstance`.