#3189 calculate "-framerate" parameter based on actual number of screenshots per second
On a slow network/browser, it may happen that Selenide took less screenshots per second than the requested FPS. In this case, the resulting video was too "fast" (faster and shorter than the reality).

When running "ffmpeg", we need to calculate "-framerate" parameter which reflects the actual number of screenshots. Then the video will be based on the actual data, and video length and speed will be close to reality.